backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
google-generativeai==0.8.3
sentence-transformers==3.2.1
chromadb==0.5.18
pandas==2.2.3
pydantic==2.9.2
